SLIM-1936 Refactoring
