+++ osgp/protocol-adapter-dlms/osgp-protocol-adapter-dlms/src/main/java/org/opensmartgridplatform/adapter/protocol/dlms/domain/commands/dlmsobjectconfig/model/Medium.java
@@ -6,9 +6,8 @@
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig;
+package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig.model;
 
-public enum RegisterUnit {
-    WH, // watt hour
-    M3  // cubic meter
+public enum Medium {
+    ABSTRACT, ELECTRICITY, GAS, COMBINED
 }
