+++ osgp/protocol-adapter-dlms/osgp-protocol-adapter-dlms/src/main/java/org/opensmartgridplatform/adapter/protocol/dlms/domain/commands/dlmsobjectconfig/model/RegisterUnit.java
@@ -6,8 +6,9 @@
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig;
+package org.opensmartgridplatform.adapter.protocol.dlms.domain.commands.dlmsobjectconfig.model;
 
-public enum Medium {
-    ABSTRACT, ELECTRICITY, GAS, COMBINED
+public enum RegisterUnit {
+    WH, // watt hour
+    M3  // cubic meter
 }
